feat(irs-api):[TRI-509] Extracted processing of shells and submodels to separate class AASHandler
