+++ irs-models/src/main/java/net/catenax/irs/component/assetadministrationshell/SubmodelDescriptor.java
@@ -11,6 +11,7 @@ package net.catenax.irs.component.assetadministrationshell;
 
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
@@ -47,5 +48,13 @@ public class SubmodelDescriptor {
      * endpoint
      */
     private List<Endpoint> endpoints;
+
+    /**
+     * @return The first value of the semanticId or null, if none is present.
+     */
+    @JsonIgnore
+    public String getAspectType() {
+        return this.getSemanticId().getValue().stream().findFirst().orElse(null);
+    }
 
 }
